refactor TracksSection
