+++ src/components/organisms/sections/TracksSection.tsx
@@ -1,65 +1,22 @@
 import OptimizedImage from "@/components/atoms/OptimizedImage";
 import ContentCard from "@/components/molecules/ContentCard";
+import FinalistCard from "@/components/molecules/track/FinalistCard";
 import SectionTitle from "@/components/molecules/SectionTitle";
+import SponsorTrack from "@/components/molecules/track/SponsorTrack";
 import { tracksData, finalistsData } from "@/data/eventData";
 import { mq } from "@/themes/settings/breakpoints";
 import { brand, neutral } from "@/themes/settings/color";
+import type { TrackId } from "@/types/finalists";
+import { getTrackFinalists, SPONSOR_TRACKS } from "@/utils/finalistsUtils";
 import { css } from "@emotion/react";
 import { useState } from "react";
 import type { FC } from "react";
-
-interface FinalistCardProps {
-  finalist: {
-    name: string;
-    image: string;
-    description: string;
-    link: string;
-    prize: string;
-  };
-}
 
 interface TrackCardProps {
   track: (typeof tracksData)[0];
   isExpanded: boolean;
   onToggle: () => void;
 }
-
-const FinalistCard: FC<FinalistCardProps> = ({ finalist }) => {
-  const hasImage = finalist.image && finalist.image.trim() !== "";
-
-  return (
-    <div css={finalistCardStyle}>
-      <div css={finalistImageWrapperStyle}>
-        {hasImage ? (
-          <OptimizedImage
-            src={finalist.image}
-            alt={finalist.name}
-            css={finalistImageStyle}
-            width={80}
-            height={80}
-          />
-        ) : (
-          <div css={finalistPlaceholderStyle}>🍕</div>
-        )}
-        <div css={finalistPrizeBadgeStyle}>{finalist.prize}</div>
-      </div>
-      <div css={finalistInfoStyle}>
-        <h4 css={finalistNameStyle}>{finalist.name}</h4>
-        <p css={finalistDescriptionStyle}>{finalist.description}</p>
-        {finalist.link && finalist.link.trim() !== "" && (
-          <a
-            href={finalist.link}
-            target="_blank"
-            rel="noopener noreferrer"
-            css={finalistLinkStyle}
-          >
-            View Project →
-          </a>
-        )}
-      </div>
-    </div>
-  );
-};
 
 const TrackCard: FC<TrackCardProps> = ({ track, isExpanded, onToggle }) => {
   const trackTitle = (
@@ -206,87 +163,42 @@ const TracksSection: FC = () => {
                 onToggle={() => toggleTrack(track.id)}
               />
 
-              {finalistsData[track.id as keyof typeof finalistsData] &&
-                finalistsData[track.id as keyof typeof finalistsData].length >
-                  0 && (
-                  <div css={finalistsContainerStyle}>
-                    <h4 css={finalistsTitleStyle}>Finalists</h4>
-                    <div css={finalistsGridStyle}>
-                      {finalistsData[
-                        track.id as keyof typeof finalistsData
-                      ].map((finalist) => (
-                        <FinalistCard
-                          key={`finalist-${track.id}-${finalist.name}`}
-                          finalist={finalist}
-                        />
-                      ))}
+              {(() => {
+                const finalists = getTrackFinalists(
+                  finalistsData,
+                  track.id as TrackId,
+                );
+                return (
+                  finalists.length > 0 && (
+                    <div css={finalistsContainerStyle}>
+                      <h4 css={finalistsTitleStyle}>Finalists</h4>
+                      <div css={finalistsGridStyle}>
+                        {finalists.map((finalist) => (
+                          <FinalistCard
+                            key={`finalist-${track.id}-${finalist.name}`}
+                            finalist={finalist}
+                          />
+                        ))}
+                      </div>
                     </div>
-                  </div>
-                )}
+                  )
+                );
+              })()}
             </div>
           ))}
 
           {/* Sponsor Track Finalists */}
           <div css={sponsorTracksContainerStyle}>
             <h3 css={sponsorTracksTitleStyle}>Sponsor Track Finalists</h3>
-
-            {/* PizzaDAO Track */}
-            {finalistsData["PizzaDAO" as keyof typeof finalistsData] &&
-              finalistsData["PizzaDAO" as keyof typeof finalistsData].length >
-                0 && (
-                <div css={sponsorTrackStyle}>
-                  <h4 css={sponsorTrackTitleStyle}>🍕 PizzaDAO Track</h4>
-                  <div css={finalistsGridStyle}>
-                    {finalistsData[
-                      "PizzaDAO" as keyof typeof finalistsData
-                    ].map((finalist) => (
-                      <FinalistCard
-                        key={`pizzadao-finalist-${finalist.name}`}
-                        finalist={finalist}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
-
-            {/* ENS Track */}
-            {finalistsData["ens" as keyof typeof finalistsData] &&
-              finalistsData["ens" as keyof typeof finalistsData].length > 0 && (
-                <div css={sponsorTrackStyle}>
-                  <h4 css={sponsorTrackTitleStyle}>🌐 ENS Track</h4>
-                  <div css={finalistsGridStyle}>
-                    {finalistsData["ens" as keyof typeof finalistsData].map(
-                      (finalist) => (
-                        <FinalistCard
-                          key={`ens-finalist-${finalist.name}`}
-                          finalist={finalist}
-                        />
-                      ),
-                    )}
-                  </div>
-                </div>
-              )}
-
-            {/* JAPAN SMART CHAIN Track */}
-            {finalistsData["JAPAN SMART CHAIN" as keyof typeof finalistsData] &&
-              finalistsData["JAPAN SMART CHAIN" as keyof typeof finalistsData]
-                .length > 0 && (
-                <div css={sponsorTrackStyle}>
-                  <h4 css={sponsorTrackTitleStyle}>
-                    🏗️ JAPAN SMART CHAIN Track
-                  </h4>
-                  <div css={finalistsGridStyle}>
-                    {finalistsData[
-                      "JAPAN SMART CHAIN" as keyof typeof finalistsData
-                    ].map((finalist) => (
-                      <FinalistCard
-                        key={`jsc-finalist-${finalist.name}`}
-                        finalist={finalist}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
+            {SPONSOR_TRACKS.map((track) => (
+              <SponsorTrack
+                key={track.id}
+                trackId={track.id}
+                title={track.title}
+                emoji={track.emoji}
+                finalists={finalistsData[track.id] || []}
+              />
+            ))}
           </div>
         </div>
       </div>
@@ -573,98 +485,6 @@ const finalistsGridStyle = css`
   }
 `;
 
-const finalistCardStyle = css`
-  background: linear-gradient(135deg, ${neutral.White} 0%, ${neutral.Grey1} 100%);
-  border-radius: 12px;
-  padding: 1.25rem;
-  border: 1px solid ${brand.Primary}20;
-  transition: all 0.3s ease;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
-  
-  &:hover {
-    transform: translateY(-4px);
-    box-shadow: 0 8px 24px rgba(0, 0, 0, 0.1);
-    border-color: ${brand.Primary}40;
-  }
-`;
-
-const finalistImageWrapperStyle = css`
-  position: relative;
-  display: flex;
-  justify-content: center;
-  margin-bottom: 1rem;
-`;
-
-const finalistImageStyle = css`
-  width: 80px;
-  height: 80px;
-  border-radius: 50%;
-  object-fit: cover;
-  border: 3px solid ${brand.Primary};
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-`;
-
-const finalistPlaceholderStyle = css`
-  width: 80px;
-  height: 80px;
-  border-radius: 50%;
-  background-color: ${neutral.Grey2};
-  border: 3px solid ${neutral.Grey4};
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 2.5rem;
-`;
-
-const finalistPrizeBadgeStyle = css`
-  position: absolute;
-  top: -12px;
-  left: 50%;
-  transform: translateX(-50%);
-  background: linear-gradient(135deg, ${brand.Primary} 0%, ${brand.Secondary} 100%);
-  color: ${neutral.White};
-  font-size: 0.75rem;
-  font-weight: 600;
-  padding: 0.25rem 0.5rem;
-  border-radius: 12px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.2);
-  white-space: nowrap;
-  z-index: 1;
-`;
-
-const finalistInfoStyle = css`
-  text-align: center;
-`;
-
-const finalistNameStyle = css`
-  color: ${neutral.Grey6};
-  font-size: 1.1rem;
-  font-weight: 700;
-  margin: 0 0 0.75rem;
-  line-height: 1.3;
-`;
-
-const finalistDescriptionStyle = css`
-  color: ${neutral.Grey5};
-  font-size: 0.9rem;
-  line-height: 1.4;
-  margin: 0 0 1rem;
-`;
-
-const finalistLinkStyle = css`
-  color: ${brand.Primary};
-  font-size: 0.9rem;
-  font-weight: 600;
-  text-decoration: none;
-  transition: color 0.2s ease;
-  
-  &:hover {
-    color: ${brand.Secondary};
-    text-decoration: underline;
-  }
-`;
-
 // Sponsor Track styles
 const sponsorTracksContainerStyle = css`
   margin-top: 3rem;
@@ -681,26 +501,6 @@ const sponsorTracksTitleStyle = css`
   
   ${mq.mobileSmall} {
     font-size: 1.5rem;
-  }
-`;
-
-const sponsorTrackStyle = css`
-  margin-bottom: 2.5rem;
-  
-  &:last-child {
-    margin-bottom: 0;
-  }
-`;
-
-const sponsorTrackTitleStyle = css`
-  color: ${brand.Secondary};
-  font-size: 1.4rem;
-  font-weight: 700;
-  margin: 0 0 1.5rem;
-  text-align: center;
-  
-  ${mq.mobileSmall} {
-    font-size: 1.2rem;
   }
 `;
